+++ relayer-cli/src/commands.rs
@@ -12,12 +12,12 @@ mod listen;
 mod query;
 mod start;
 mod tx;
-mod utils;
+mod v0;
 mod version;
 
 use self::{
-    config::ConfigCmd, keys::KeysCmd, light::LightCmd, listen::ListenCmd, query::QueryCmd,
-    start::StartCmd, tx::TxCmd, version::VersionCmd,
+    config::ConfigCmd, light::LightCmd, listen::ListenCmd, query::QueryCmd, start::StartCmd,
+    tx::TxCmd, v0::V0Cmd, version::VersionCmd, keys::KeysCmd
 };
 
 use crate::config::Config;
@@ -33,6 +33,10 @@ pub enum CliCmd {
     /// The `help` subcommand
     #[options(help = "get usage information")]
     Help(Help<Self>),
+
+    /// The `v0` subcommand
+    #[options(help = "start the v0 relayer")]
+    V0(V0Cmd),
 
     /// The `start` subcommand
     #[options(help = "start the relayer")]